refactor: Change path of marching cube related files to 'Geometry'
